Move custom stuff out of framework.h
